requirements.txt
django-oscar==1.5.*
pillow
psycopg2-binary
sorl-thumbnail
pycountry
django-compressor
django-widget-tweaks
squareup
latimes-statestyle
